DOP-6219: Breadcrumb fix
